\name{ICCov-package}
\alias{ICCov-package}
\alias{ICCov}
\docType{package}
\title{
\packageTitle{ICCov}
}
\description{
\packageDescription{ICCov}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ICCov}
\packageIndices{ICCov}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ICCov}

Maintainer: \packageMaintainer{ICCov}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
