% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICCov_lm.R
\name{ICCov_lm}
\alias{ICCov_lm}
\title{Fit semiparametric regression models with interval-censored covariates}
\usage{
ICCov_lm(
  formula,
  covariate_function,
  data,
  hazard_formula = NULL,
  tolerance = 1e-04,
  pl_tolerance = 1e-04,
  max_iter = 500,
  stepsize = NULL
)
}
\arguments{
\item{formula}{A object of class `"formula"`: a symbolic description of the model to be fit. See `Details` for more information.}

\item{covariate_function}{A `"function"` object which specifies how the intermediate event time and response are linked. See `Details` for examples.}

\item{data}{An optional object of class `"formula"` specifying how the intermediate event is linked to the covariates. If left absent, then the intermediate event will be modeled as dependent on all covariates included to the `formula` argument.}

\item{tolerance}{A scalar value for the convergence tolerance of the EM algorithm.}

\item{pl_tolerance}{A scalar value for the convergence tolerance of the profile likelihood method used for variance estimation.}

\item{stepsize}{A scalar value for the stepsize used in the profile likelihood method. By default, 1/(5*n^1/2) is used.}
}
\description{
Fit semiparametric regression models with interval-censored covariates
}
