% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICCov_lm.R
\name{timedResponse}
\alias{timedResponse}
\title{Creates a "timedResponse" class variable for input into ICCov regression models}
\usage{
timedResponse(response, measurement_time)
}
\arguments{
\item{response}{A numeric response vector of interest}

\item{measurement_time}{A numeric vector containing the measurement time for each element in response}
}
\description{
Creates a "timedResponse" class variable for input into ICCov regression models
}
